Use Transform Space more and fixes... When changing the transform space from the default Local Space, update the default type and values as well.
Fix a minor error in the get_action_constraints() function, when in object mode.

Move the src and limit from the operator.
Having them in the running operator made it difficult to update the other values alongside transform space.
The only issue with this would be that you can't run the operator multiple times because the local class would be in use, though I don't know how that would be possible anyway.